+++ Realm/RLMSyncManager.mm
@@ -36,10 +36,12 @@
 
 using namespace realm;
 
+// NEXT-MAJOR: All the code associated to the logger from sync manager should be removed.
 using Level = realm::util::Logger::Level;
 
 namespace {
-
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wdeprecated-declarations"
 Level levelForSyncLogLevel(RLMSyncLogLevel logLevel) {
     switch (logLevel) {
         case RLMSyncLogLevelOff:    return Level::off;
@@ -69,6 +71,7 @@ RLMSyncLogLevel logLevelForLevel(Level logLevel) {
     }
     REALM_UNREACHABLE();    // Unrecognized log level.
 }
+#pragma clang diagnostic pop
 
 #pragma mark - Loggers