+++ Realm/Tests/RealmTests.mm
@@ -24,6 +24,7 @@
 #import "RLMRealm_Dynamic.h"
 #import "RLMRealm_Private.hpp"
 #import "RLMSchema_Private.h"
+#import "RLMLogger_Private.h"
 
 #import <mach/mach_init.h>
 #import <mach/vm_map.h>
@@ -2937,5 +2938,97 @@
     XCTAssertEqual(error.code, RLMErrorAlreadyOpen);
     XCTAssertTrue([NSFileManager.defaultManager fileExistsAtPath:config.fileURL.path]);
 }
+@end
 
+@interface RLMLoggerTests : RLMTestCase
+@property (nonatomic, strong) RLMLogger *logger;
+@end
+
+@implementation RLMLoggerTests
+- (void)setUp {
+    _logger = RLMLogger.defaultLogger;
+}
+- (void)tearDown {
+    RLMLogger.defaultLogger = _logger;
+}
+- (void)testSetDefaultLogLevel {
+    __block NSString *logs = @"";
+    RLMLogger *logger = [[RLMLogger alloc] initWithLevel:RLMLogLevelAll logFunction:^(RLMLogLevel level, NSString *message) {
+        NSString *newLogs = [logs stringByAppendingFormat:@" %@ %lu %@", [NSDate date], level, message];
+        logs = newLogs;
+    }];
+    RLMLogger.defaultLogger = logger;
+
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertEqual([RLMLogger defaultLogger].level, RLMLogLevelAll);
+    XCTAssertTrue([logs containsString:@"5 DB:"]); // Detail
+    XCTAssertTrue([logs containsString:@"7 DB:"]); // Trace
+
+    logs = @"";
+    logger.level = RLMLogLevelDetail;
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertEqual([RLMLogger defaultLogger].level, RLMLogLevelDetail);
+    XCTAssertTrue([logs containsString:@"5 DB:"]); // Detail
+    XCTAssertFalse([logs containsString:@"7 DB:"]); // Trace
+}
+
+- (void)testDefaultLogger {
+    __block NSString *logs = @"";
+    RLMLogger *logger = [[RLMLogger alloc] initWithLevel:RLMLogLevelOff
+                                             logFunction:^(RLMLogLevel level, NSString *message) {
+        NSString *newLogs = [logs stringByAppendingFormat:@" %@ %lu %@", [NSDate date], level, message];
+        logs = newLogs;
+    }];
+    RLMLogger.defaultLogger = logger;
+    XCTAssertEqual(RLMLogger.defaultLogger.level, RLMLogLevelOff);
+
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertTrue([logs length] == 0);
+
+    // Test LogLevel Detail
+    logger.level = RLMLogLevelDetail;
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertTrue([logs length] >= 0);
+    XCTAssertTrue([logs containsString:@"5 DB:"]); // Detail
+    XCTAssertFalse([logs containsString:@"7 DB:"]); // Trace
+
+    // Test LogLevel All
+    logger.level = RLMLogLevelAll;
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertTrue([logs length] >= 0);
+    XCTAssertTrue([logs containsString:@"5 DB:"]); // Detail
+    XCTAssertTrue([logs containsString:@"7 DB:"]); // Trace
+
+    logs = @"";
+    // Init Custom Logger
+    RLMLogger.defaultLogger = [[RLMLogger alloc] initWithLevel:RLMLogLevelDebug
+                                                   logFunction:^(RLMLogLevel level, NSString * message) {
+        NSString *newLogs = [logs stringByAppendingFormat:@" %@ %lu %@", [NSDate date], level, message];
+        logs = newLogs;
+    }];
+
+    XCTAssertEqual(RLMLogger.defaultLogger.level, RLMLogLevelDebug);
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertTrue([logs containsString:@"5 DB:"]); // Detail
+    XCTAssertFalse([logs containsString:@"7 DB:"]); // Trace
+}
+
+- (void)testCustomLoggerLogMessage {
+    __block NSString *logs = @"";
+    RLMLogger *logger = [[RLMLogger alloc] initWithLevel:RLMLogLevelInfo
+                                             logFunction:^(RLMLogLevel level, NSString * message) {
+        NSString *newLogs = [logs stringByAppendingFormat:@" %@ %lu %@", [NSDate date], level, message];
+        logs = newLogs;
+
+    }];
+    RLMLogger.defaultLogger = logger;
+
+    @autoreleasepool { [RLMRealm defaultRealm]; }
+    XCTAssertTrue([logs length] >= 0);
+
+    [logger logWithLevel:RLMLogLevelInfo message:@"%@ IMPORTANT INFO %i", @"TEST:", 0];
+    [logger logWithLevel:RLMLogLevelTrace message:@"IMPORTANT TRACE"];
+    XCTAssertTrue([logs containsString:@"TEST: IMPORTANT INFO 0"]); // Detail
+    XCTAssertFalse([logs containsString:@"IMPORTANT TRACE"]); // Trace
+}
 @end