+++ RealmSwift/Sync.swift
@@ -230,6 +230,7 @@ extension AppError {
 
  - see: `RLMSyncLogLevel`
  */
+@available(*, deprecated)
 public typealias SyncLogLevel = RLMSyncLogLevel
 
 /**