+++ include/Realm/RLMLogger_Private.h
@@ -1,0 +1,1 @@
+../../Realm/RLMLogger_Private.h